+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Cash App parser: added new export format.
 - Coinbase parser: added "Cash to Savings", "Savings to Cash", "Interest payout" and "Retail Simple Dust" transaction types.
 - Exodus parser: added new export format. ([#467](https://github.com/BittyTax/BittyTax/issues/467))
-- Coinbase parser: added "Subscription" transaction type.
+- Coinbase parser: added "Subscription", "Prime Deposit", "Transfer", "Retail Staking Transfer" and "Retail Unstaking Transfer" transaction types.
 - Conversion tool: added parser for Coinbase Prime exchange.
 ### Changed
 - Price tool: CoinDesk API deprecated.